+++ CRM/Core/BAO/SchemaHandler.php
@@ -1,9 +1,9 @@
 <?php
 /*
  +--------------------------------------------------------------------+
- | CiviCRM version 4.7                                                |
+ | CiviCRM version 5                                                  |
  +--------------------------------------------------------------------+
- | Copyright CiviCRM LLC (c) 2004-2016                                |
+ | Copyright CiviCRM LLC (c) 2004-2019                                |
  +--------------------------------------------------------------------+
  | This file is a part of CiviCRM.                                    |
  |                                                                    |
@@ -28,7 +28,7 @@
 /**
  *
  * @package CRM
- * @copyright CiviCRM LLC (c) 2004-2016
+ * @copyright CiviCRM LLC (c) 2004-2019
  */
 
 /**
@@ -67,8 +67,7 @@ class CRM_Core_BAO_SchemaHandler {
   public static function createTable(&$params) {
     $sql = self::buildTableSQL($params);
     // do not i18n-rewrite
-    $dao = CRM_Core_DAO::executeQuery($sql, array(), TRUE, NULL, FALSE, FALSE);
-    $dao->free();
+    CRM_Core_DAO::executeQuery($sql, [], TRUE, NULL, FALSE, FALSE);
 
     $config = CRM_Core_Config::singleton();
     if ($config->logging) {
@@ -189,14 +188,15 @@ class CRM_Core_BAO_SchemaHandler {
 
     //create index only for searchable fields during ADD,
     //create index only if field is become searchable during MODIFY,
-    //drop index only if field is no more searchable and index was exist.
+    //drop index only if field is no longer searchable and it does not reference
+    //a forgein key (and indexExist is true)
     if (!empty($params['searchable']) && !$indexExist) {
       $sql .= $separator;
       $sql .= str_repeat(' ', 8);
       $sql .= $prefix;
       $sql .= "INDEX_{$params['name']} ( {$params['name']} )";
     }
-    elseif (empty($params['searchable']) && $indexExist) {
+    elseif (empty($params['searchable']) && empty($params['fk_table_name']) && $indexExist) {
       $sql .= $separator;
       $sql .= str_repeat(' ', 8);
       $sql .= "DROP INDEX INDEX_{$params['name']}";
@@ -252,15 +252,13 @@ class CRM_Core_BAO_SchemaHandler {
 ALTER TABLE {$tableName}
       DROP FOREIGN KEY `FK_{$fkName}`;";
 
-    $dao = CRM_Core_DAO::executeQuery($dropFKSql);
-    $dao->free();
+    CRM_Core_DAO::executeQuery($dropFKSql);
 
     $addFKSql = "
 ALTER TABLE {$tableName}
       ADD CONSTRAINT `FK_{$fkName}` FOREIGN KEY (`entity_id`) REFERENCES {$fkTableName} (`id`) ON DELETE CASCADE;";
     // CRM-7007: do not i18n-rewrite this query
-    $dao = CRM_Core_DAO::executeQuery($addFKSql, array(), TRUE, NULL, FALSE, FALSE);
-    $dao->free();
+    CRM_Core_DAO::executeQuery($addFKSql, [], TRUE, NULL, FALSE, FALSE);
 
     return TRUE;
   }
@@ -335,8 +333,7 @@ ALTER TABLE {$tableName}
     }
 
     // CRM-7007: do not i18n-rewrite this query
-    $dao = CRM_Core_DAO::executeQuery($sql, array(), TRUE, NULL, FALSE, FALSE);
-    $dao->free();
+    CRM_Core_DAO::executeQuery($sql, [], TRUE, NULL, FALSE, FALSE);
 
     $config = CRM_Core_Config::singleton();
     if ($config->logging) {
@@ -345,7 +342,7 @@ ALTER TABLE {$tableName}
       // Are there any modifies we DON'T was to call this function for (& shouldn't it be clever enough to cope?)
       if ($params['operation'] == 'add' || $params['operation'] == 'modify') {
         $logging = new CRM_Logging_Schema();
-        $logging->fixSchemaDifferencesFor($params['table_name'], array(trim($prefix) => array($params['name'])), FALSE);
+        $logging->fixSchemaDifferencesFor($params['table_name'], [trim($prefix) => [$params['name']]], FALSE);
       }
     }
 
@@ -357,36 +354,37 @@ ALTER TABLE {$tableName}
   }
 
   /**
-   * Delete a CiviCRM-table
+   * Delete a CiviCRM-table.
    *
    * @param string $tableName
    *   Name of the table to be created.
    */
   public static function dropTable($tableName) {
     $sql = "DROP TABLE $tableName";
-    $dao = CRM_Core_DAO::executeQuery($sql);
+    CRM_Core_DAO::executeQuery($sql);
   }
 
   /**
    * @param string $tableName
    * @param string $columnName
    * @param bool $l18n
+   * @param bool $isUpgradeMode
    *
    */
-  public static function dropColumn($tableName, $columnName, $l18n = FALSE) {
+  public static function dropColumn($tableName, $columnName, $l18n = FALSE, $isUpgradeMode = FALSE) {
     if (self::checkIfFieldExists($tableName, $columnName)) {
       $sql = "ALTER TABLE $tableName DROP COLUMN $columnName";
       if ($l18n) {
         CRM_Core_DAO::executeQuery($sql);
       }
       else {
-        CRM_Core_DAO::executeQuery($sql, array(), TRUE, NULL, FALSE, FALSE);
+        CRM_Core_DAO::executeQuery($sql, [], TRUE, NULL, FALSE, FALSE);
       }
       $domain = new CRM_Core_DAO_Domain();
       $domain->find(TRUE);
       if ($domain->locales) {
         $locales = explode(CRM_Core_DAO::VALUE_SEPARATOR, $domain->locales);
-        CRM_Core_I18n_Schema::rebuildMultilingualSchema($locales, NULL);
+        CRM_Core_I18n_Schema::rebuildMultilingualSchema($locales, NULL, $isUpgradeMode);
       }
     }
   }
@@ -406,7 +404,7 @@ ADD INDEX `FK_{$tableName}_entity_id` ( `entity_id` )";
 DROP INDEX `FK_{$tableName}_entity_id` ,
 ADD UNIQUE INDEX `unique_entity_id` ( `entity_id` )";
     }
-    $dao = CRM_Core_DAO::executeQuery($sql);
+    CRM_Core_DAO::executeQuery($sql);
   }
 
   /**
@@ -428,10 +426,10 @@ ADD UNIQUE INDEX `unique_entity_id` ( `entity_id` )";
    * @todo add support for length & multilingual on combined keys.
    *
    * @param string $createIndexPrefix
-   * @param array $substrLenghts
+   * @param array $substrLengths
    */
-  public static function createIndexes($tables, $createIndexPrefix = 'index', $substrLenghts = array()) {
-    $queries = array();
+  public static function createIndexes($tables, $createIndexPrefix = 'index', $substrLengths = []) {
+    $queries = [];
     $domain = new CRM_Core_DAO_Domain();
     $domain->find(TRUE);
     $locales = explode(CRM_Core_DAO::VALUE_SEPARATOR, $domain->locales);
@@ -446,7 +444,7 @@ ADD UNIQUE INDEX `unique_entity_id` ( `entity_id` )";
       $query = "SHOW INDEX FROM $table";
       $dao = CRM_Core_DAO::executeQuery($query);
 
-      $currentIndexes = array();
+      $currentIndexes = [];
       while ($dao->fetch()) {
         $currentIndexes[] = $dao->Key_name;
       }
@@ -464,16 +462,16 @@ ADD UNIQUE INDEX `unique_entity_id` ( `entity_id` )";
         else {
           // handle indices over substrings, CRM-6245
           // $lengthName is appended to index name, $lengthSize is the field size modifier
-          $lengthName = isset($substrLenghts[$table][$fieldName]) ? "_{$substrLenghts[$table][$fieldName]}" : '';
-          $lengthSize = isset($substrLenghts[$table][$fieldName]) ? "({$substrLenghts[$table][$fieldName]})" : '';
+          $lengthName = isset($substrLengths[$table][$fieldName]) ? "_{$substrLengths[$table][$fieldName]}" : '';
+          $lengthSize = isset($substrLengths[$table][$fieldName]) ? "({$substrLengths[$table][$fieldName]})" : '';
         }
 
-        $names = array(
+        $names = [
           "index_{$fieldName}{$lengthName}",
           "FK_{$table}_{$fieldName}{$lengthName}",
           "UI_{$fieldName}{$lengthName}",
           "{$createIndexPrefix}_{$fieldName}{$lengthName}",
-        );
+        ];
 
         // skip to the next $field if one of the above $names exists; handle multilingual for CRM-4126
         foreach ($names as $name) {
@@ -507,6 +505,31 @@ ADD UNIQUE INDEX `unique_entity_id` ( `entity_id` )";
   }
 
   /**
+   * Get indexes for tables
+   * @param array $tables
+   *   array of table names to find indexes for
+   *
+   * @return array('tableName' => array('index1', 'index2'))
+   */
+  public static function getIndexes($tables) {
+    $indexes = [];
+    foreach ($tables as $table) {
+      $query = "SHOW INDEX FROM $table";
+      $dao = CRM_Core_DAO::executeQuery($query);
+
+      $tableIndexes = [];
+      while ($dao->fetch()) {
+        $tableIndexes[$dao->Key_name]['name'] = $dao->Key_name;
+        $tableIndexes[$dao->Key_name]['field'][] = $dao->Column_name .
+         ($dao->Sub_part ? '(' . $dao->Sub_part . ')' : '');
+        $tableIndexes[$dao->Key_name]['unique'] = ($dao->Non_unique == 0 ? 1 : 0);
+      }
+      $indexes[$table] = $tableIndexes;
+    }
+    return $indexes;
+  }
+
+  /**
    * Drop an index if one by that name exists.
    *
    * @param string $tableName
@@ -533,10 +556,10 @@ UPDATE civicrm_custom_field
 SET    text_length = %1
 WHERE  id = %2
 ";
-    $params = array(
-      1 => array($length, 'Integer'),
-      2 => array($customFieldID, 'Integer'),
-    );
+    $params = [
+      1 => [$length, 'Integer'],
+      2 => [$customFieldID, 'Integer'],
+    ];
     CRM_Core_DAO::executeQuery($sql, $params);
 
     $sql = "
@@ -566,11 +589,11 @@ MODIFY      {$columnName} varchar( $length )
     }
     else {
       CRM_Core_Error::fatal(ts('Could Not Find Custom Field Details for %1, %2, %3',
-        array(
+        [
           1 => $tableName,
           2 => $columnName,
           3 => $customFieldID,
-        )
+        ]
       ));
     }
   }
@@ -581,12 +604,12 @@ MODIFY      {$columnName} varchar( $length )
    * @param string $tableName
    * @param array $indexName
    *
-   * @return \CRM_Core_DAO|object
+   * @return bool
    */
   public static function checkIfIndexExists($tableName, $indexName) {
     $result = CRM_Core_DAO::executeQuery(
       "SHOW INDEX FROM $tableName WHERE key_name = %1 AND seq_in_index = 1",
-      array(1 => array($indexName, 'String'))
+      [1 => [$indexName, 'String']]
     );
     if ($result->fetch()) {
       return TRUE;
@@ -595,32 +618,29 @@ MODIFY      {$columnName} varchar( $length )
   }
 
   /**
-   * Check if the table has a specified column
+   * Check if the table has a specified column.
    *
    * @param string $tableName
    * @param string $columnName
+   * @param bool $i18nRewrite
+   *   Whether to rewrite the query on multilingual setups.
    *
-   * @return \CRM_Core_DAO|object
+   * @return bool
    */
-  public static function checkIfFieldExists($tableName, $columnName) {
-    $result = CRM_Core_DAO::executeQuery(
-      "SHOW COLUMNS FROM $tableName LIKE %1",
-      array(1 => array($columnName, 'String'))
-    );
-    if ($result->fetch()) {
-      return TRUE;
-    }
-    return FALSE;
+  public static function checkIfFieldExists($tableName, $columnName, $i18nRewrite = TRUE) {
+    $query = "SHOW COLUMNS FROM $tableName LIKE '%1'";
+    $dao = CRM_Core_DAO::executeQuery($query, [1 => [$columnName, 'Alphanumeric']], TRUE, NULL, FALSE, $i18nRewrite);
+    $result = $dao->fetch() ? TRUE : FALSE;
+    return $result;
   }
 
   /**
-   * Remove a foreign key from a table if it exists
-   *
-   * @param $table_name
-   * @param $constraint_name
+   * Check if a foreign key Exists
+   * @param string $table_name
+   * @param string $constraint_name
+   * @return bool TRUE if FK is found
    */
-  public static function safeRemoveFK($table_name, $constraint_name) {
-
+  public static function checkFKExists($table_name, $constraint_name) {
     $config = CRM_Core_Config::singleton();
     $dbUf = DB::parseDSN($config->dsn);
     $query = "
@@ -630,18 +650,133 @@ MODIFY      {$columnName} varchar( $length )
       AND CONSTRAINT_NAME = %3
       AND CONSTRAINT_TYPE = 'FOREIGN KEY'
     ";
-    $params = array(
-      1 => array($dbUf['database'], 'String'),
-      2 => array($table_name, 'String'),
-      3 => array($constraint_name, 'String'),
-    );
+    $params = [
+      1 => [$dbUf['database'], 'String'],
+      2 => [$table_name, 'String'],
+      3 => [$constraint_name, 'String'],
+    ];
     $dao = CRM_Core_DAO::executeQuery($query, $params);
 
     if ($dao->fetch()) {
-      CRM_Core_DAO::executeQuery("ALTER TABLE {$table_name} DROP FOREIGN KEY {$constraint_name}", array());
       return TRUE;
     }
     return FALSE;
+  }
+
+  /**
+   * Remove a foreign key from a table if it exists.
+   *
+   * @param $table_name
+   * @param $constraint_name
+   *
+   * @return bool
+   */
+  public static function safeRemoveFK($table_name, $constraint_name) {
+    if (self::checkFKExists($table_name, $constraint_name)) {
+      CRM_Core_DAO::executeQuery("ALTER TABLE {$table_name} DROP FOREIGN KEY {$constraint_name}", []);
+      return TRUE;
+    }
+    return FALSE;
+  }
+
+  /**
+   * Add index signature hash to DAO file calculation.
+   *
+   * @param string $table table name
+   * @param array $indices index array spec
+   */
+  public static function addIndexSignature($table, &$indices) {
+    foreach ($indices as $indexName => $index) {
+      $indices[$indexName]['sig'] = $table . "::" .
+        (array_key_exists('unique', $index) ? $index['unique'] : 0) . "::" .
+        implode("::", $index['field']);
+    }
+  }
+
+  /**
+   * Compare the indices specified in the XML files with those in the DB.
+   *
+   * @param bool $dropFalseIndices
+   *  If set - this function deletes false indices present in the DB which mismatches the expected
+   *  values of xml file so that civi re-creates them with correct values using createMissingIndices() function.
+   *
+   * @return array
+   *   index specifications
+   */
+  public static function getMissingIndices($dropFalseIndices = FALSE) {
+    $requiredSigs = $existingSigs = [];
+    // Get the indices defined (originally) in the xml files
+    $requiredIndices = CRM_Core_DAO_AllCoreTables::indices();
+    $reqSigs = [];
+    foreach ($requiredIndices as $table => $indices) {
+      $reqSigs[] = CRM_Utils_Array::collect('sig', $indices);
+    }
+    CRM_Utils_Array::flatten($reqSigs, $requiredSigs);
+
+    // Get the indices in the database
+    $existingIndices = CRM_Core_BAO_SchemaHandler::getIndexes(array_keys($requiredIndices));
+    $extSigs = [];
+    foreach ($existingIndices as $table => $indices) {
+      CRM_Core_BAO_SchemaHandler::addIndexSignature($table, $indices);
+      $extSigs[] = CRM_Utils_Array::collect('sig', $indices);
+    }
+    CRM_Utils_Array::flatten($extSigs, $existingSigs);
+
+    // Compare
+    $missingSigs = array_diff($requiredSigs, $existingSigs);
+
+    //CRM-20774 - Drop index key which exist in db but the value varies.
+    $existingKeySigs = array_intersect_key($missingSigs, $existingSigs);
+    if ($dropFalseIndices && !empty($existingKeySigs)) {
+      foreach ($existingKeySigs as $sig) {
+        $sigParts = explode('::', $sig);
+        foreach ($requiredIndices[$sigParts[0]] as $index) {
+          if ($index['sig'] == $sig && !empty($index['name'])) {
+            self::dropIndexIfExists($sigParts[0], $index['name']);
+            continue;
+          }
+        }
+      }
+    }
+
+    // Get missing indices
+    $missingIndices = [];
+    foreach ($missingSigs as $sig) {
+      $sigParts = explode('::', $sig);
+      if (array_key_exists($sigParts[0], $requiredIndices)) {
+        foreach ($requiredIndices[$sigParts[0]] as $index) {
+          if ($index['sig'] == $sig) {
+            $missingIndices[$sigParts[0]][] = $index;
+            continue;
+          }
+        }
+      }
+    }
+    return $missingIndices;
+  }
+
+  /**
+   * Create missing indices.
+   *
+   * @param array $missingIndices as returned by getMissingIndices()
+   */
+  public static function createMissingIndices($missingIndices) {
+    $queries = [];
+    foreach ($missingIndices as $table => $indexList) {
+      foreach ($indexList as $index) {
+        $queries[] = "CREATE " .
+        (array_key_exists('unique', $index) && $index['unique'] ? 'UNIQUE ' : '') .
+        "INDEX {$index['name']} ON {$table} (" .
+          implode(", ", $index['field']) .
+        ")";
+      }
+    }
+
+    /* FIXME potential problem if index name already exists, so check before creating */
+    $dao = new CRM_Core_DAO();
+    foreach ($queries as $query) {
+      $dao->query($query, FALSE);
+    }
   }
 
 }
